code update / result upload

+++ src/kalman_filter.cpp
@@ -56,14 +56,14 @@ void KalmanFilter::UpdateEKF(const VectorXd &z) {
     * update the state by using Extended Kalman Filter equations
   */
     // H matrix * (px, py, vx, vy)T => (rho, phi, rho dot)
-    double px = x_[0];
-    double py = x_[1];
-    double vx = x_[2];
-    double vy = x_[3];
+    float px = x_[0];
+    float py = x_[1];
+    float vx = x_[2];
+    float vy = x_[3];
 
-    double rho = sqrt( px*px + py*py );
-    double phi = atan2( py, px );
-    double rho_dot = ( px*vx + py*vy )/rho;
+    float rho = sqrt( px*px + py*py );
+    float phi = atan2( py, px );
+    float rho_dot = ( px*vx + py*vy )/rho;
 
     // y = z - hxprime
     VectorXd hxprime = VectorXd(3);
